requirements.txt
unstructured==0.11.6
python-magic-bin==0.4.14
langchain==0.0.353
charset-normalizer==3.3.2
cohere==4.39
flask==3.0.2
Flask-Cors==4.0.0
pinecone-client==3.0.2
PyJWT==2.8.0
pdf2image==1.17.0
pillow==10.2.0
pdfminer==20191125